requirements.txt
lxml
tqdm
tensorflow==2.3.0rc0
absl-py
easydict
matplotlib
pillow
numpy==1.22.0
opencv-python==4.5.5.62
urllib3==1.26.8